+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 -------------------------------------------------------------------------------------------------------------
 
-# 5.7.15 (2021-10-18)
+# 5.7.15 (2021-10-19)
 
 ### 🐣新特性
 * 【db     】     Db.quietSetAutoCommit增加判空（issue#I4D75B@Gitee）
@@ -18,6 +18,7 @@
 * 【poi    】     修复ExcelWriter多余调试信息导致的问题（issue#1884@Github）
 * 【poi    】     修复TemporalAccessorUtil.toInstant使用DateTimeFormatter导致问题（issue#1891@Github）
 * 【poi    】     修复sheet.getRow(y)为null导致的问题（issue#1893@Github）
+* 【cache  】     修复LRUCache线程安全问题（issue#1895@Github）
 
 -------------------------------------------------------------------------------------------------------------
