+++ hutool-cache/src/main/java/cn/hutool/cache/impl/LFUCache.java
@@ -15,7 +15,7 @@ import java.util.Iterator;
  * @param <K> 键类型
  * @param <V> 值类型
  */
-public class LFUCache<K, V> extends AbstractCache<K, V> {
+public class LFUCache<K, V> extends StampedCache<K, V> {
 	private static final long serialVersionUID = 1L;
 
 	/**
